requirements.txt
fastapi
uvicorn
tensorflow
numpy
pandas
pydantic
python-multipart
scikit-learn
openai
python-dotenv